+++ src/runtime/src/core/range.rs
@@ -1,4 +1,4 @@
-use crate::{external_error, Value, ValueIterator, ValueMap};
+use crate::{external_error, IntRange, Value, ValueIterator, ValueMap};
 
 pub fn make_module() -> ValueMap {
     use Value::*;
@@ -13,6 +13,24 @@ pub fn make_module() -> ValueMap {
     result.add_fn("end", |vm, args| match vm.get_args(args) {
         [Range(r)] => Ok(Number(r.end.into())),
         _ => external_error!("range.end: Expected range as argument"),
+    });
+
+    result.add_fn("expanded", |vm, args| match vm.get_args(args) {
+        [Range(r), Number(n)] => {
+            let n = isize::from(n);
+            if r.is_ascending() {
+                Ok(Range(IntRange {
+                    start: r.start - n,
+                    end: r.end + n,
+                }))
+            } else {
+                Ok(Range(IntRange {
+                    start: r.start + n,
+                    end: r.end - n,
+                }))
+            }
+        }
+        _ => external_error!("range.expanded: Expected range and number as arguments"),
     });
 
     result.add_fn("iter", |vm, args| match vm.get_args(args) {