+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+- Core Ops
+  - `range.expanded`
+
 
 ## [0.5.0] 2020.12.17